requirements.txt
fastapi
uvicorn
motor
python-jose
passlib
python-multipart